// src/generated/models/Pm_projectapprovalrequestsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_projectapprovalrequestspm_approvalstage = {
  0: 'BusinessCaseApproval',
  1: 'BudgetRelease',
  2: 'ChangeApproval',
  3: 'RemediationApproval',
  4: 'PipelineApproval'
} as const;
export type Pm_projectapprovalrequestspm_approvalstage = keyof typeof Pm_projectapprovalrequestspm_approvalstage;
export const Pm_projectapprovalrequestspm_decisionstatus = {
  0: 'Approved',
  1: 'Pending',
  2: 'Rejected'
} as const;
export type Pm_projectapprovalrequestspm_decisionstatus = keyof typeof Pm_projectapprovalrequestspm_decisionstatus;
export const Pm_projectapprovalrequestspm_entitytype = {
  0: 'Project',
  1: 'ChangeRequest'
} as const;
export type Pm_projectapprovalrequestspm_entitytype = keyof typeof Pm_projectapprovalrequestspm_entitytype;
export const Pm_projectapprovalrequestspm_prioritylevel = {
  0: 'High',
  1: 'Medium',
  2: 'Critical'
} as const;
export type Pm_projectapprovalrequestspm_prioritylevel = keyof typeof Pm_projectapprovalrequestspm_prioritylevel;
export const Pm_projectapprovalrequestsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_projectapprovalrequestsstatecode = keyof typeof Pm_projectapprovalrequestsstatecode;
export const Pm_projectapprovalrequestsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_projectapprovalrequestsstatuscode = keyof typeof Pm_projectapprovalrequestsstatuscode;

export interface Pm_projectapprovalrequestsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pm_approvalstage?: Pm_projectapprovalrequestspm_approvalstage;
  pm_approvername?: string;
  pm_decisiondate?: string;
  pm_decisionnotes?: string;
  pm_decisionstatus?: Pm_projectapprovalrequestspm_decisionstatus;
  pm_duedate?: string;
  pm_entityid?: string;
  pm_entitytype?: Pm_projectapprovalrequestspm_entitytype;
  pm_prioritylevel?: Pm_projectapprovalrequestspm_prioritylevel;
  pm_projectapprovalrequestid: string;
  pm_requestorname?: string;
  pm_requesttitle?: string;
  statecode: Pm_projectapprovalrequestsstatecode;
  statuscode?: Pm_projectapprovalrequestsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_projectapprovalrequests extends Pm_projectapprovalrequestsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_approvalstagename?: string;
  pm_decisionstatusname?: string;
  pm_entitytypename?: string;
  pm_prioritylevelname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}